pyproject.toml
[project]
name = "blogmark"
version = "0.1.0"
description = "Personal RAG-powered knowledge base for blog posts"
readme = "README.md"
authors = [
    { name = "Runtian (Rachel) Li", email = "Rachel_0619@hotmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.5.0",
    "lxml>=4.9.0",
    "html5lib>=1.1",
    "jupyter>=1.1.1",
    "notebook>=7.4.4",
    "ipykernel>=6.29.5",
    "trafilatura>=1.6.1",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "qdrant-client[fastembed]>=1.14.2",
    "llama-index>=0.12.46",
    "llama-index-vector-stores-qdrant>=0.6.1",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-embeddings-ollama>=0.6.0",
    "llama-index-node-parser-topic>=0.2.0",
    "spacy>=3.8.7",
    "en-core-web-md",
    "streamlit>=1.46.1",
]

[project.scripts]
blogmark = "blogmark:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
en-core-web-md = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl" }
